[v0.90][WP-14] Split tooling command test responsibilities (Closes #2111)
